CentOS installation beta release

+++ var/www/html/bin/ipxe-cleand.sh
@@ -1,14 +1,16 @@
 mkdir -p /mnt/{nfs,rootfs/boot}
 mount iPXE_Server_IP:/pxelinux.cfg /mnt/nfs
 
-# Bind NIC device to ethX
-mount /dev/sda6 /mnt/rootfs/boot
-echo set linux_append="net.ifnames=0 biosdevname=0" > /mnt/rootfs/boot/grub.cfg
-
 # Remove PXE configuration
 mac=`cat /root/installation_nic_mac`
 macd=01-`echo ${mac,,} | tr ':' '-'`
 cd /mnt/nfs
 rm $macd
 
-reboot
+# Change Network Interface naming to tradition
+if grep -i coreos /etc/*-release; then
+    mount /dev/sda6 /mnt/rootfs/boot
+    echo set linux_append="net.ifnames=0 biosdevname=0" > /mnt/rootfs/boot/grub.cfg
+    umount -l /mnt/rootfs/boot
+    reboot
+fi
